requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
langgraph
pydantic
python-dotenv
pydantic_settings
pytest
psycopg2-binary 
chromadb 
google-genai